backend/django_core/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.10
django-redis==5.4.0

# Environment Management
django-environ==0.11.2
python-decouple==3.8

# Authentication & Authorization
django-allauth==0.57.0
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# Async Task Processing (for AI agent communication)
celery==5.3.4
redis==5.0.1

# File Handling
Pillow==10.0.1
python-magic==0.4.27

# HTTP Client for FastAPI communication
httpx==0.25.1
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.26.4

# Monitoring & Logging
sentry-sdk==1.38.0

# Development Tools
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0

# Utilities
uuid==1.30
python-dateutil==2.8.2
pytz==2023.3
